Registries: add program registry + generate bundle; enforce via guard
