+++ engine/src/phases/phase4/util.ts
@@ -1,0 +1,20 @@
+import type { ExerciseSignature } from "../../substitution/types.js";
+
+export function uniqueStable(ids: string[]): string[] {
+  const seen = new Set<string>();
+  const out: string[] = [];
+  for (const id of ids) {
+    const s = String(id ?? "").trim();
+    if (!s) continue;
+    if (seen.has(s)) continue;
+    seen.add(s);
+    out.push(s);
+  }
+  return out;
+}
+
+export function pick(entries: Record<string, ExerciseSignature>, id: string): ExerciseSignature {
+  const ex = entries?.[id];
+  if (!ex) throw new Error(`Missing exercise ${id}`);
+  return ex;
+}
