+++ test/fixtures/golden/expected/phase3_precedence_banned_over_available.json
@@ -37,7 +37,8 @@
       "loaded_registries": [
         "activity",
         "movement",
-        "exercise"
+        "exercise",
+        "program"
       ],
       "notes": [
         "PHASE_3: registries loaded",
